src: Move redux folder outside of component/update import statements
