+++ src/components/add_a_book/Add_a_Book.js
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux';
 
 import { v4 as uuidv4 } from 'uuid';
 
-import { addBook } from '../redux/books/books';
+import { addBook } from '../../redux/books/books';
 
 const AddBook = () => {
   const [title, setTitle] = useState('');
